Revert "Merge branch '11.x' into 2865710-dependencies-from-only"
This reverts commit 4e202c332d768ef5c50b6e58a01b402799a9f599, reversing
changes made to 08a410edb45bf49d2bc890b28655e80662b2f8d2.

I shouldn't have tried to merge 11.x into this MR since the target branch can't be changed.  The MR will have to be closed, but I need to be able to see its changes easily.

+++ core/modules/path/src/Plugin/Field/FieldType/PathItem.php
@@ -3,25 +3,24 @@
 namespace Drupal\path\Plugin\Field\FieldType;
 
 use Drupal\Component\Utility\Random;
-use Drupal\Core\Field\Attribute\FieldType;
 use Drupal\Core\Field\FieldDefinitionInterface;
-use Drupal\Core\Field\FieldItemBase;
 use Drupal\Core\Field\FieldStorageDefinitionInterface;
-use Drupal\Core\StringTranslation\TranslatableMarkup;
+use Drupal\Core\Field\FieldItemBase;
 use Drupal\Core\TypedData\DataDefinition;
 
 /**
  * Defines the 'path' entity field type.
+ *
+ * @FieldType(
+ *   id = "path",
+ *   label = @Translation("Path"),
+ *   description = @Translation("An entity field containing a path alias and related data."),
+ *   no_ui = TRUE,
+ *   default_widget = "path",
+ *   list_class = "\Drupal\path\Plugin\Field\FieldType\PathFieldItemList",
+ *   constraints = {"PathAlias" = {}},
+ * )
  */
-#[FieldType(
-  id: "path",
-  label: new TranslatableMarkup("Path"),
-  description: new TranslatableMarkup("An entity field containing a path alias and related data."),
-  default_widget: "path",
-  no_ui: TRUE,
-  list_class: PathFieldItemList::class,
-  constraints: ["PathAlias" => []],
-)]
 class PathItem extends FieldItemBase {
 
   /**
@@ -48,20 +47,15 @@ class PathItem extends FieldItemBase {
    * {@inheritdoc}
    */
   public function isEmpty() {
-    $alias = $this->get('alias')->getValue();
-    $pid = $this->get('pid')->getValue();
-    $langcode = $this->get('langcode')->getValue();
-
-    return ($alias === NULL || $alias === '') && ($pid === NULL || $pid === '') && ($langcode === NULL || $langcode === '');
+    return ($this->alias === NULL || $this->alias === '') && ($this->pid === NULL || $this->pid === '') && ($this->langcode === NULL || $this->langcode === '');
   }
 
   /**
    * {@inheritdoc}
    */
   public function preSave() {
-    $alias = $this->get('alias')->getValue();
-    if ($alias !== NULL) {
-      $this->set('alias', trim($alias));
+    if ($this->alias !== NULL) {
+      $this->alias = trim($this->alias);
     }
   }
 
@@ -71,38 +65,35 @@ class PathItem extends FieldItemBase {
   public function postSave($update) {
     $path_alias_storage = \Drupal::entityTypeManager()->getStorage('path_alias');
     $entity = $this->getEntity();
-    $alias = $this->get('alias')->getValue();
-    $pid = $this->get('pid')->getValue();
-    $langcode = $this->get('langcode')->getValue();
 
     // If specified, rely on the langcode property for the language, so that the
     // existing language of an alias can be kept. That could for example be
     // unspecified even if the field/entity has a specific langcode.
-    $alias_langcode = ($langcode && $pid) ? $langcode : $this->getLangcode();
+    $alias_langcode = ($this->langcode && $this->pid) ? $this->langcode : $this->getLangcode();
 
     // If we have an alias, we need to create or update a path alias entity.
-    if ($alias) {
-      if (!$update || !$pid) {
+    if ($this->alias) {
+      if (!$update || !$this->pid) {
         $path_alias = $path_alias_storage->create([
           'path' => '/' . $entity->toUrl()->getInternalPath(),
-          'alias' => $alias,
+          'alias' => $this->alias,
           'langcode' => $alias_langcode,
         ]);
         $path_alias->save();
-        $this->set('pid', $path_alias->id());
+        $this->pid = $path_alias->id();
       }
-      else {
-        $path_alias = $path_alias_storage->load($pid);
+      elseif ($this->pid) {
+        $path_alias = $path_alias_storage->load($this->pid);
 
-        if ($alias != $path_alias->getAlias()) {
-          $path_alias->setAlias($alias);
+        if ($this->alias != $path_alias->getAlias()) {
+          $path_alias->setAlias($this->alias);
           $path_alias->save();
         }
       }
     }
-    elseif ($pid) {
+    elseif ($this->pid && !$this->alias) {
       // Otherwise, delete the old alias if the user erased it.
-      $path_alias = $path_alias_storage->load($pid);
+      $path_alias = $path_alias_storage->load($this->pid);
       if ($entity->isDefaultRevision()) {
         $path_alias_storage->delete([$path_alias]);
       }
